+++ Linux/setup-linux.sh
@@ -9,7 +9,7 @@ source "$DOTFILES/Linux/Root/dot-functions.sh"
 
 message "🔔 Environment:" "Locations being used for this install of Dotfiles"
 bullet "DOTFILES = $DOTFILES"
-bullet "Run location = ${0:a:h}"
+bullet "Installing from locatiion = ${0:a:h}"
 bullet "git config --global user.name = \"$(git config --get user.name)\""
 bullet "git config --global user.email = \"$(git config --get user.email)\""
 
@@ -17,11 +17,11 @@ bullet "git config --global user.email = \"$(git config --get user.email)\""
 # TODO: abort if git, cc, or other needed tools are not available
 
 
-# ==============================================================================
+# ========================================================================
 # Require `bash` as the default on Linux, and set the default shell if needed
 if [ $SHELL != "/bin/bash" ]; then
   chsh -s /bin/bash
-  echo "❌ Script has set default shell to BASH, now re-run setup.sh"
+  echo "Script has set default shell to BASH, now re-run setup.sh"
   exit 0
 fi
 
@@ -36,20 +36,32 @@ chmod -R 777 ${DOTFILES}/*    2> /dev/null
 find $DOTFILES -name "*.sh" -type f -print0 | xargs -0 chmod 755
 
 
-# TODO: replace this with the XDG setup now in Mac for #HOME/.local/...
-# ==============================================================================
-# Create $HOME/Bin folder in which to put local code repositorities
-if [[ -d "$HOME/Bin" ]]; then
-  bullet "$HOME/Bin exists. Added to the PATH for user content"
+
+# ========================================================================
+# Create XDG:  https://specifications.freedesktop.org/basedir/latest/
+export XDG_BIN_HOME=$HOME/.local/bin
+export XDG_CONFIG_HOME=$HOME/.config
+export XDG_DATA_HOME=$HOME/.local/share
+export XDG_STATE_HOME=$HOME/.local/state
+export XDG_CACHE_HOME=$HOME/.local/cache
+
+mkdir -p $XDG_BIN_HOME    2> /dev/null
+mkdir -p $XDG_CONFIG_HOME 2> /dev/null
+mkdir -p $XDG_DATA_HOME    2> /dev/null
+mkdir -p $XDG_STATE_HOME   2> /dev/null
+mkdir -p $XDG_CACHE_HOME   2> /dev/null
+
+if [[ -d "$HOME/.local/bin" ]]; then
+  bullet "$HOME/.local/bin exists. Added to the PATH for user content"
 else
-  mkdir -p $HOME/Bin
-  chown -R $USER $HOME/Bin
-  chmod 744 $HOME/Bin
-  message "📂 Created $HOME/Bin and added it to PATH for your code"
+  chown -R "$USER":admin $XDG_BIN_HOME
+  chmod 0700 $XDG_BIN_HOME
+  message "📂 Created $HOME/.local/bin and (will) add it to PATH"
 fi
 
 
-# ==============================================================================
+
+# ========================================================================
 message "✅ Installing root dotfiles" "Overwriting existing versions of these files"
 cp $DOTFILES/Linux/Root/dot-bashrc.sh $HOME/.bashrc
 cp $DOTFILES/Linux/Root/dot-aliases.sh $HOME/.aliases
@@ -65,12 +77,12 @@ cp $DOTFILES/Common/Root/dot-gitignore $HOME/.gitignore
 # Register gitignore and other git stuff
 git config --global core.excludesfile ~/.gitignore
 
-# ==============================================================================
-message "✅ Copy scripts to PATH" "Using $HOME/Bin for user scripts"
-cp $DOTFILES/Linux/Bin/* $HOME/Bin
+# ========================================================================
+message "✅ Copy scripts to PATH" "Using $XDG_BIN_HOME for user scripts"
+cp $DOTFILES/Linux/Bin/* $XDG_BIN_HOME
 
 
-# ==============================================================================
+# ========================================================================
 message "✅ Setup defaults" "Adding paths and variables to .bashrc for bash"
 echo " " >> ~/.bashrc
 echo "# Add global DOTFILES pointing Dotfiles install folder" >> ~/.bashrc
@@ -82,12 +94,12 @@ cp $HOME/.bashrc $HOME/.bash_profile
 cp $HOME/.bashrc $HOME/.profile
 
 
-# ==============================================================================
+# ========================================================================
 # On Linux copy the ./Common/Eza/* theme to ~/.config/eza/theme.yml
 # TODO: set this up on Linux
 
 
-# ==============================================================================
+# ========================================================================
 # Check if the "~/local.sh" file exists, and if not, copy  the stub version to user home
 if [[ -f "$HOME/local.sh" ]]; then
   bullet "~/local.sh exists - delete then re-run to reset from the template"
@@ -96,10 +108,11 @@ else
   cp $DOTFILES/Linux/Root/local.sh $HOME/local.sh
 fi
 
-# ==============================================================================
+# ========================================================================
 echo
-message "🎉 Success!" "Restart Terminal and run setup-brew.sh and setup-ruby.sh"
-echo
+message "🎉 Success!" "Restart Terminal to see the changes."
+
 
 ## end of file.
+echo
 exit 0