+++ Mac/Root/dot-aliases.sh
@@ -57,8 +57,9 @@ then
   # Make standard 'ls' look prettier, with colors and sorting
   # alias ls='ls -FG'
 else
-  ## Aliasing 'ls' commands to use 'eza'
+  # Aliasing 'ls' commands to use 'eza' - default is in "grid" mode
   alias ll='eza --long --sort=Name --git --git-repos -I "Icon?" --group-directories-first --no-quotes --no-permissions --no-user --grid'
+  alias lll='eza --long --sort=Name --git --git-repos -I "Icon?" --group-directories-first --no-quotes --no-permissions --no-user'
   alias lls='eza --sort=Name --git --git-repos -I "Icon?" --group-directories-first --no-quotes'
   alias lla='eza -a --long --sort=Name --git -I "Icon?" -I ".DS_Store" --group-directories-first'
   alias llx='eza -a --long --sort=Name --git -I "Icon?" -I ".DS_Store" --group-directories-first -@ -Z'